Share country-only billing address assertions
Committed-By-Agent: codex
